Added modular generation feature

+++ src/Support/Console/Commands/MakeResourceModel.php
@@ -14,7 +14,7 @@ class MakeResourceModel extends Command
      *
      * @var string
      */
-    protected $signature = 'support:resource:model {resources*}';
+    protected $signature = 'support:resource:model {resources*} {--module=*}';
 
     /**
      * The console command description.
@@ -46,21 +46,47 @@ class MakeResourceModel extends Command
         $ds = DIRECTORY_SEPARATOR;
         $resources = $this->argument('resources');
         $resources = !is_array($resources) ? [$resources] : $resources;
+        $module = $this->option('module');
+        $directory = app_path();
+
+        if ($module) {
+            $module = is_array($module) ? $module[0] : $module;
+            $moduleDirectory = config('support.module.directory') . DIRECTORY_SEPARATOR . ucfirst($module);
+
+            $this->namespace = $this->getAppNamespace() . ucfirst($module) . '\\' . str_replace($ds, '\\', config('support.module.models'));
+            $directory = config('support.module.directory') . DIRECTORY_SEPARATOR . ucfirst($module) . $ds . config('support.module.models');
+            
+            if (!file_exists($moduleDirectory)) {
+                $this->error("The requested module '$moduleDirectory' does not exists!");
+                return;
+            }
+
+            if(config('support.namespace')) {
+                $this->namespace = config('support.namespace');
+            }
+        } else {
+            $this->namespace = rtrim($this->namespace, "\\");
+        }
         
         foreach ($resources as $resource) {
-            $stub = file_get_contents(__DIR__ . $ds . '..' . $ds . '..' . $ds . '..' . $ds . '..' . $ds . 'resources' . $ds . 'Stubs' . $ds . 'Model.stub');
-            $stub = str_replace([
-                        '{{namespace}}',   '{{resource}}'], 
-                        [rtrim($this->namespace, "\\"),  str_singular(ucfirst($resource))],
-                        $stub);
 
-            $directory = app_path();
+            $stub = file_get_contents(__DIR__ . $ds . '..' . $ds . '..' . $ds . '..' . $ds . '..' . $ds . 'resources' . $ds . 'Stubs' . $ds . (empty($module) ? 'Standard' : 'Modular') . $ds . 'Model.stub');
+            $stub = str_replace(
+                        ['{{namespace}}',  '{{resource}}'], 
+                        [$this->namespace,  str_singular(ucfirst($resource))],
+                        $stub);
 
             if (!file_exists($directory)) {
                 mkdir($directory, 0777, true);
             }
 
             $filename = $directory . $ds . str_singular(ucfirst($resource)) . '.php';
+            
+            if (file_exists($filename)) {
+                $this->info($filename . ' already exists!');
+                return;
+            }
+
             file_put_contents($filename, $stub);
             $this->info($filename . ' created!');
         }
